split gpu and cpu test=develop
