+++ python/paddle/fluid/tests/unittests/CMakeLists.txt
@@ -284,9 +284,6 @@ py_test_modules(test_parallel_executor_crf MODULES test_parallel_executor_crf)
 py_test_modules(test_parallel_executor_crf_auto_growth MODULES test_parallel_executor_crf_auto_growth ENVS FLAGS_allocator_strategy=auto_growth)
 py_test_modules(test_parallel_executor_fetch_feed MODULES test_parallel_executor_fetch_feed)
 set_tests_properties(test_parallel_executor_fetch_feed PROPERTIES TIMEOUT 450)
-set_tests_properties(test_parallel_executor_seresnext_base PROPERTIES TIMEOUT 300)
-set_tests_properties(test_parallel_executor_seresnext_with_reduce PROPERTIES TIMEOUT 250)
-set_tests_properties(test_parallel_executor_seresnext_with_fuse_all_reduce PROPERTIES TIMEOUT 250)
 py_test_modules(test_parallel_executor_transformer MODULES test_parallel_executor_transformer)
 py_test_modules(test_parallel_executor_transformer_auto_growth MODULES test_parallel_executor_transformer_auto_growth ENVS FLAGS_allocator_strategy=auto_growth)
 py_test_modules(test_layers MODULES test_layers ENVS FLAGS_cudnn_deterministic=1)
@@ -295,9 +292,9 @@ if(NOT WIN32)
 endif()
 
 if(CMAKE_BUILD_TYPE STREQUAL "Debug")
-    set_tests_properties(test_parallel_executor_seresnext_base PROPERTIES TIMEOUT 900)
-    set_tests_properties(test_parallel_executor_seresnext_with_reduce PROPERTIES TIMEOUT 740)
-    set_tests_properties(test_parallel_executor_seresnext_with_fuse_all_reduce PROPERTIES TIMEOUT 450)
+    set_tests_properties(test_parallel_executor_seresnext_base_cpu PROPERTIES TIMEOUT 900)
+    set_tests_properties(test_parallel_executor_seresnext_with_reduce_cpu PROPERTIES TIMEOUT 740)
+    set_tests_properties(test_parallel_executor_seresnext_with_fuse_all_reduce_cpu PROPERTIES TIMEOUT 450)
 endif()
 
 if (WITH_NGRAPH)
@@ -309,8 +306,8 @@ if (WITH_MKLDNN)
 endif()
 
 set_tests_properties(test_parallel_executor_test_while_train test_parallel_executor_mnist
-        test_parallel_executor_seresnext_base test_parallel_executor_seresnext_with_reduce
-        test_parallel_executor_seresnext_with_fuse_all_reduce
+        test_parallel_executor_seresnext_base_gpu test_parallel_executor_seresnext_with_reduce_gpu
+        test_parallel_executor_seresnext_with_fuse_all_reduce_gpu
         test_parallel_executor_crf test_sync_batch_norm_op
         test_parallel_executor_crf_auto_growth test_buffer_shared_memory_reuse_pass_and_fuse_optimization_op_pass
         test_buffer_shared_memory_reuse_pass PROPERTIES LABELS "RUN_TYPE=DIST")
