+++ python/paddle/fluid/tests/unittests/test_parallel_executor_seresnext_base_cpu.py
@@ -13,29 +13,24 @@
 # limitations under the License.
 
 from __future__ import print_function
-import paddle.fluid as fluid
-fluid.core._set_fuse_parameter_group_size(3)
-fluid.core._set_fuse_parameter_memory_size(131072)
-
 import unittest
 import seresnext_net
 from seresnext_test_base import TestResnetBase
 from functools import partial
 
 
-class TestResnetWithFuseAllReduce(TestResnetBase):
-    def test_seresnext_with_fused_all_reduce(self):
-        # NOTE(zcd): In order to make the program faster,
-        # this unit test remove drop_out and batch_norm.
-        # rm_drop_out=True,
-        # rm_bn=True,
+class TestResnetCPU(TestResnetBase):
+    def test_seresnext_with_learning_rate_decay(self):
+        # NOTE(zcd): This test is compare the result of use parallel_executor
+        # and executor, and the result of drop_out op and batch_norm op in
+        # this two executor have diff, so the two ops should be removed
+        # from the model.
         check_func = partial(
             self.check_network_convergence,
             optimizer=seresnext_net.optimizer,
-            fuse_all_reduce_ops=True)
-        self._compare_result_with_origin_model(check_func, use_cuda=False)
+            use_parallel_executor=False)
         self._compare_result_with_origin_model(
-            check_func, use_cuda=True, delta2=1e-2)
+            check_func, use_cuda=False, compare_seperately=False, delta2=1e-3)
 
 
 if __name__ == '__main__':
