+++ python/paddle/fluid/tests/unittests/test_parallel_executor_seresnext_with_reduce_gpu.py
@@ -1,0 +1,28 @@
+# Copyright (c) 2019 PaddlePaddle Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from __future__ import print_function
+import unittest
+from test_parallel_executor_seresnext_with_reduce_cpu import TestResnetWithReduceBase
+
+
+class TestResnetWithReduceGPU(TestResnetWithReduceBase):
+    # TODO(zcd): temporally disable reduce_and_allreduce test because of the random failure.
+    @unittest.skip("should fix this later.")
+    def test_seresnext_with_reduce(self):
+        self._compare_reduce_and_allreduce(use_cuda=True, delta2=1e-2)
+
+
+if __name__ == '__main__':
+    unittest.main()
